+++ examples/demo.c
@@ -124,6 +124,7 @@ static void load_default_document(void) {
                     TINTA_OPTION_KEYBOARD_NAVIGATION |
                     TINTA_OPTION_MOUSE_ZOOM |
                     TINTA_OPTION_CODE_COPY_BUTTON |
+                    TINTA_OPTION_DOCUMENT_COPY_BUTTON |
                     TINTA_OPTION_LOCAL_IMAGES |
                     TINTA_OPTION_REMOTE_IMAGES |
                     TINTA_OPTION_OPEN_UNHANDLED_LINKS;